requirements.txt
geopandas==1.0.0
pandas==2.3.0
numpy==1.23.5
torch==2.7.1
torchvision==0.22.1
timm==0.3.2  # need for MAE
python-dotenv==1.1.0
openai==1.90.0
glob2==0.7
fsspec==2025.5.1
pyogrio==0.11.0
sentence_transformers==4.1.0
transformers==4.41.0
qdrant-client==1.14.3
deep_translator==1.11.4
geopy==2.4.1
cdifflib==1.2.9
